chatbot/requirements.txt
apiai==1.2.3
atomicwrites==1.4.0
attrs==20.2.0
certifi==2020.6.20
cffi==1.14.1
click==7.1.2
colorama==0.4.3
cryptography==3.0
DAWG-Python==0.7.2
decorator==4.4.2
docopt==0.6.2
flaky==3.7.0
fuzzywuzzy==0.18.0
iniconfig==1.0.1
joblib==0.16.0
more-itertools==8.5.0
nltk==3.5
numpy==1.19.1
packaging==20.4
Pillow==8.1.0
pluggy==0.13.1
pony==0.7.14
psycopg2==2.8.6
ptbtest==1.3
py==1.9.0
pycparser==2.20
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
pyparsing==2.4.7
pytest==6.2.1
python-dateutil==2.8.1
python-Levenshtein==0.12.0
python-telegram-bot==12.8
regex==2020.7.14
six==1.15.0
toml==0.10.1
tornado==6.0.4
tqdm==4.48.2
